refactor: use badgescontainer for serverinfo badges
