+++ src/context/intermediate/popovers/ServerInfo.tsx
@@ -9,6 +9,7 @@ import styles from "./ServerInfo.module.scss";
 import { Localizer, Text } from "preact-i18n";
 import { useEffect, useState } from "preact/hooks";
 
+import BadgesContainer from "../../../components/common/BadgesContainer";
 import ServerIcon from "../../../components/common/ServerIcon";
 import Tooltip from "../../../components/common/Tooltip";
 import UserIcon from "../../../components/common/user/UserIcon";
@@ -135,58 +136,14 @@ export const ServerInfo = observer(
                         </>
 
                         {server.flags ? (
-                            <div className={styles.category}>Badges</div>
-                        ) : undefined}
-                        {server.flags && server.flags & 1 ? (
-                            <Tooltip
-                                content={
-                                    <Text id="app.special.server-badges.official" />
-                                }
-                                placement={"bottom-start"}>
-                                <svg width="20" height="20">
-                                    <image
-                                        xlinkHref="/assets/badges/verified.svg"
-                                        height="20"
-                                        width="20"
-                                    />
-                                    <image
-                                        xlinkHref="/assets/badges/revolt_r.svg"
-                                        height="15"
-                                        width="15"
-                                        x="2"
-                                        y="3"
-                                        style={
-                                            "justify-content: center; align-items: center; filter: brightness(0);"
-                                        }
-                                    />
-                                </svg>
-                            </Tooltip>
-                        ) : undefined}
-                        {server.flags && server.flags & 2 ? (
-                            <Tooltip
-                                content={
-                                    <Text id="app.special.server-badges.verified" />
-                                }
-                                placement={"bottom-start"}>
-                                <svg width="20" height="20">
-                                    <image
-                                        xlinkHref="/assets/badges/verified.svg"
-                                        height="20"
-                                        width="20"
-                                    />
-                                    <foreignObject
-                                        x="2"
-                                        y="2"
-                                        width="15"
-                                        height="15">
-                                        <Check
-                                            size={15}
-                                            color="black"
-                                            strokeWidth={8}
-                                        />
-                                    </foreignObject>
-                                </svg>
-                            </Tooltip>
+                            <>
+                                <div className={styles.category}>Badges</div>
+                                <BadgesContainer
+                                    badges={server.flags}
+                                    id={server._id}
+                                    type="server"
+                                />
+                            </>
                         ) : undefined}
                         <div className={styles.category}>description</div>
                         {server.description ? (
